// src/user/user.enum.ts
export enum GenderEnum {
    MALE = 'MALE',
    FEMALE = 'FEMALE'
}

export enum LangEnum {
    EN = 'EN',
    AR = 'AR'
}

export enum DeviceEnum {
    DESKTOP = 'DESKTOP',
    IOS = 'IOS',
    ANDROID = 'ANDROID'
}

export enum UserRoleType {
    ADMIN = 'ADMIN',
    EDITOR = 'EDITOR',
    USER = 'USER'
}

export enum UserVerificationCodeUseCaseEnum {
    EMAIL_VERIFICATION = 'EMAIL_VERIFICATION',
    PASSWORD_RESET = 'PASSWORD_RESET',
    SOCIAL_REGISTER = 'SOCIAL_REGISTER',
}